Changed openmama root header approach
There are some files which OpenMAMA has no business using and their use
is discouraged in modern OpenMAMA based code anyway. With that in mind,
this patch moves these files to the "wombat" include subdirectory
instead so that they are still available for use, but are no longer in a
location which is likely to cause conflicts with other libraries.